pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wholesale-clustering-mlops"
version = "0.1.0"
description = "POC MLOps pour la segmentation de clients grossistes"
authors = [
    {name = "Alexandre Mathias Donnat"},
    {name = "Lounis Hamroun"},
    {name = "Leo Ivars"},
    {name = "Omar Fekih-Hassen"},
    {name = "Anne Faury"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # --- Data & Machine Learning ---
    "ucimlrepo>=0.0.7",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scipy>=1.11.0",
    # --- API (Serving) ---
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",

    # --- WebApp (Frontend) ---
    "gradio>=4.0.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",

    # --- Tracking & Registry ---
    "mlflow>=2.10.0,<3.0.0",
    "boto3>=1.34.0",
    "PyMySQL>=1.1.0",

    # --- Utils ---
    "python-dotenv>=1.0.0",

    # --- Tests ---
    "pytest>=8.0.0",
]

[project.optional-dependencies]
dev = [
    # --- Tests & CI/CD ---
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "pytest-cov>=5.0.0",

]

[tool.pytest.ini_options]
addopts = "-v --disable-warnings"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src", "api", "webapp"]